requirements.txt
discord.py
boto3
python-weather